refactor(ip): migrate dashboard.order to component
