+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order/order.component.js
@@ -1,0 +1,11 @@
+import controller from './order.controller';
+import template from './order.html';
+
+export default {
+  bindings: {
+    goBack: '<',
+    serviceNameToOrder: '<',
+  },
+  controller,
+  template,
+};
