+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order/order.html
@@ -1,0 +1,189 @@
+<div
+    data-wizard
+    data-wizard-bread-crumb="true"
+    data-wizard-on-cancel="resumeOrder"
+    data-wizard-on-finish="redirectToPaymentPage"
+    data-wizard-title=":: 'ip_additional_order' | translate"
+>
+    <div
+        data-wizard-step
+        data-wizard-step-on-load="loadServices"
+        data-wizard-step-valid="$ctrl.model.selectedService"
+    >
+        <oui-message
+            class="d-block mb-4"
+            data-type="info"
+            data-ng-if="$ctrl.user.isEnterprise"
+        >
+            <span data-translate="core_user_enterprise_retail_prices"></span>
+        </oui-message>
+
+        <div data-ng-show="!$ctrl.loading.services && $ctrl.services.length">
+            <p data-translate="ip_order_step1_question"></p>
+            <form>
+                <div class="form-group">
+                    <label
+                        for="service"
+                        class="control-label"
+                        data-translate="ip_table_header_service"
+                    >
+                    </label>
+                    <select
+                        id="service"
+                        class="form-control"
+                        data-ng-model="$ctrl.model.selectedService"
+                        data-ng-options="service as service.displayName group by $ctrl.getServiceTypeLabel(service.type) for service in $ctrl.services | orderBy:'displayName' track by service.serviceName"
+                    >
+                        <option
+                            value=""
+                            data-ng-disabled="true"
+                            data-translate="global_select"
+                        >
+                        </option>
+                    </select>
+                </div>
+            </form>
+        </div>
+
+        <div
+            data-ng-show="!$ctrl.loading.services && !$ctrl.services.length"
+            data-translate="ip_order_noservices"
+        ></div>
+
+        <div data-ng-show="$ctrl.loading.services" class="text-center">
+            <oui-spinner></oui-spinner>
+        </div>
+    </div>
+
+    <div
+        data-wizard-step
+        data-wizard-step-on-load="loadIpOffers"
+        data-wizard-step-valid="$ctrl.isOfferFormValid()"
+    >
+        <form data-ng-show="!$ctrl.loading.ipOffers">
+            <div class="form-group">
+                <label
+                    for="offer"
+                    class="control-label"
+                    data-translate="ip_agora_table_header_offer"
+                >
+                </label>
+                <select
+                    id="offer"
+                    class="form-control"
+                    data-ng-model="$ctrl.model.params.selectedOffer"
+                    data-ng-change="$ctrl.onSelectedOfferChange()"
+                    data-ng-disabled="!$ctrl.ipOffers.length"
+                    data-ng-options="offer as (offer.productShortName + ' - ' + offer.price.text) group by offer.productRegion for offer in $ctrl.ipOffers | orderBy:'price.value' track by offer.planCode+offer.maximumQuantity"
+                >
+                    <option
+                        value=""
+                        data-ng-disabled="true"
+                        data-translate="global_select"
+                    >
+                    </option>
+                </select>
+                <p
+                    data-ng-if="!$ctrl.ipOffers.length"
+                    data-translate="ip_order_quota_full"
+                ></p>
+            </div>
+
+            <div data-ng-show="$ctrl.model.params.selectedOffer">
+                <div
+                    class="form-group"
+                    data-ng-show="!$ctrl.model.params.selectedOffer.isIpBlockOffer"
+                >
+                    <label
+                        for="quantity"
+                        class="control-label"
+                        data-translate="ip_agora_table_header_quantity"
+                    >
+                    </label>
+                    <select
+                        id="quantity"
+                        class="form-control"
+                        data-ng-model="$ctrl.model.params.selectedQuantity"
+                        data-ng-options="quantity for quantity in $ctrl.model.params.selectedOffer.quantities track by quantity"
+                    >
+                        <option
+                            value=""
+                            data-ng-disabled="true"
+                            data-translate="global_select"
+                        >
+                        </option>
+                    </select>
+                </div>
+
+                <div
+                    class="form-group"
+                    data-ng-show="$ctrl.model.params.selectedOffer.countries.length > 1"
+                >
+                    <label
+                        for="country"
+                        class="control-label"
+                        data-translate="ip_agora_table_header_country"
+                    >
+                    </label>
+                    <select
+                        id="country"
+                        class="form-control"
+                        data-ng-model="$ctrl.model.params.selectedCountry"
+                        data-ng-options="country as country.description for country in $ctrl.model.params.selectedOffer.countries | orderBy:'description':false:stringLocaleSensitiveComparator track by country.code"
+                    >
+                        <option
+                            value=""
+                            data-ng-disabled="true"
+                            data-translate="global_select"
+                        >
+                        </option>
+                    </select>
+                </div>
+
+                <div
+                    class="form-group"
+                    data-ng-show="$ctrl.model.params.selectedOffer.isIpBlockOffer"
+                >
+                    <label
+                        for="organisation"
+                        class="control-label"
+                        data-translate="ip_agora_table_header_organisation"
+                    >
+                    </label>
+                    <select
+                        id="organisation"
+                        class="form-control"
+                        data-ng-model="$ctrl.model.params.selectedOrganisation"
+                        data-ng-options="organisation.organisationId for organisation in $ctrl.organisations | orderBy:'organisationId' track by organisation.organisationId"
+                    >
+                        <option
+                            value=""
+                            data-ng-disabled="true"
+                            data-translate="global_select"
+                        >
+                        </option>
+                    </select>
+                    <button
+                        type="button"
+                        class="oui-link"
+                        data-translate="ip_order_no_organisation_2"
+                        data-ng-click="$ctrl.redirectToOrganisationPage()"
+                    ></button>
+                </div>
+            </div>
+        </form>
+
+        <div data-ng-show="$ctrl.loading.ipOffers" class="text-center">
+            <oui-spinner></oui-spinner>
+        </div>
+    </div>
+
+    <div data-wizard-step>
+        <p
+            class="alert alert-info"
+            data-translate="ip_order_final_txt"
+            role="alert"
+        ></p>
+        <p data-translate="ip_order_final_txt2"></p>
+    </div>
+</div>
